Move `coin_store_protocol.py` to `chia/full_node`
